requirements.txt
streamlit
google-generativeai
openai
Pillow